[lldb][test] Combine libstdc++ and libc++ std::string tests into generic test
This combines the libc++ and libstdc++ test cases. The main difference was that the libstdcpp tests had additional tests for references/pointers to std::string. I moved those over.

Split out from https://github.com/llvm/llvm-project/pull/146740
