+++ lldb/test/API/functionalities/data-formatter/data-formatter-stl/generic/string/Makefile
@@ -1,6 +1,3 @@
 CXX_SOURCES := main.cpp
 
-CFLAGS_EXTRAS := -O0
-USE_LIBSTDCPP := 1
-
 include Makefile.rules
