--- lldb/test/API/functionalities/data-formatter/data-formatter-stl/libcxx/string/Makefile
@@ -1,6 +1,0 @@
-CXX_SOURCES := main.cpp
-
-USE_LIBCPP := 1
-
-CXXFLAGS_EXTRAS := -std=c++11 -O0
-include Makefile.rules
